pyproject.toml
[project]
name = "salaryprediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "mlflow<3.0",
    "nltk>=3.9.2",
    "numpy>=2.3.5",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "pandantic>=1.0.1",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "tokenizers>=0.22.1",
    "torch>=2.9.1",
    "torchmetrics>=1.8.2",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
